+++ autotransform/batcher/base.py
@@ -10,18 +10,21 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import Any, List, Mapping, Optional, TypedDict
+from typing import Any, Generic, List, Mapping, TypedDict, TypeVar
 
 from autotransform.batcher.type import BatcherType
 from autotransform.common.cachedfile import CachedFile
 
+TParams = TypeVar("TParams", bound=Mapping[str, Any])
+
 
 class BatchMetadata(TypedDict):
-    """An object containing the metadata associated with a batch."""
+    """An object containing the metadata associated with a batch. May
+    include other fields as desired by Batcher and Repo/Transformer/etc.
+    All fields should be JSON encodable.
+    """
 
     title: str
-    summary: Optional[str]
-    tests: Optional[str]
 
 
 class Batch(TypedDict):
@@ -38,21 +41,21 @@ class BatcherBundle(TypedDict):
     type: BatcherType
 
 
-class Batcher(ABC):
+class Batcher(Generic[TParams], ABC):
     """The base for Batcher components.
 
     Attributes:
-        params (Mapping[str, Any]): The paramaters that control operation of the Batcher.
+        params (TParams): The paramaters that control operation of the Batcher.
             Should be defined using a TypedDict in subclasses
     """
 
-    params: Mapping[str, Any]
+    params: TParams
 
-    def __init__(self, params: Mapping[str, Any]):
+    def __init__(self, params: TParams):
         """A simple constructor.
 
         Args:
-            params (Mapping[str, Any]): The paramaters used to set up the Batcher
+            params (TParams): The paramaters used to set up the Batcher
         """
         self.params = params