+++ autotransform/repo/github.py
@@ -73,6 +73,19 @@ class GithubRepo(GitRepo):
             return Github(Config.get_github_username(), Config.get_github_password(), base_url=url)
         return Github(Config.get_github_username(), Config.get_github_password())
 
+    def _get_body(self, batch: Batch) -> str:
+        # pylint: disable=no-self-use
+        summary = batch["metadata"].get("summary", "")
+        tests = batch["metadata"].get("tests", "")
+
+        return f"""
+            SUMMARY
+            {summary}
+            
+            TESTS
+            {tests}
+        """
+
     def submit(self, batch: Batch) -> None:
         """Performs the normal submit for a git repo then submits a pull request
         against the provided Github repo.
@@ -81,8 +94,6 @@ class GithubRepo(GitRepo):
             batch (Batch): The Batch for which the changes were made
         """
         title = batch["metadata"]["title"]
-        summary = batch["metadata"].get("summary", "")
-        tests = batch["metadata"].get("tests", "")
 
         self.commit(batch["metadata"])
 
@@ -91,13 +102,7 @@ class GithubRepo(GitRepo):
         self.local_repo.git.push(remote.name, "-u", commit_branch)
 
         github_repo = self.github.get_repo(self.params["full_github_name"])
-        body = f"""
-            SUMMARY
-            {summary}
-            
-            TESTS
-            {tests}
-        """
+        body = self._get_body(batch)
         github_repo.create_pull(
             title=title, body=body, base=self.active_branch.name, head=commit_branch
         )